algs4: Ex 1.3.47: Add Queue/Stack/Steque classes
Queue, Stack and Steque are thin wrappers around the
CircularLinkedList class that expose certain of its methods
to provide the expected interface for their abstract data
types.

(Steque exposes all of the CircularLinkedList’s methods
because the list is singly-linked, so removal from the rear
of the list is not implemented.)

All three of the classes support `catenate(other)`, which is
the goal of this exercise, but this is itself just a thin
wrapper around the implementation of `catenate` on the
CircularLinkedList class itself, and the behaviour of the
catenation is identical (other’s items go at the back).